Moving Quiz class into an appropriate quiz.py. Still useless models
